Remove AppWidget resize frame when all apps is showing
Normally, when we drag and drop a resizable App Widget, a
resize frame will appear around it. You can change resize
frame to resize App Widget. Try quickly press all apps
button when you drop an App Widget and it's resize frame
has not appeared. You will see that the resize frame is on
the top of AppsCustomizeTabHost which is not expected. This
is what this commit solves.

The resize Runnable will run when the drop animation is
finished but will not do necessary check. This commit add
a condition to check if Workspace is in normal state
(State.NORMAL). If not, just ignore resize frame.

Change-Id: I73ce43f64fa6ba82bf70befa3b1d9d6d99c1a491