Update patch set 1
Patch Set 1:

isBreak is only used for work breaks. The actual search happens using prefix patch. So unless the letter you are searching for is in the middle, it should still work.
What is the app name and what query are you using to find the app?

Patch-set: 1
Reviewer: Gerrit User 1060398 <1060398@85c56323-6fa9-3386-8a01-6480fb634889>