Move the app to /system_ext
As mentioned in go/partitions-in-r, we plan to forbid hidden API in
product partition.
So this app should be moved to /system_ext because it uses hidden API.
It is not permanent, so they can be moved back to /product due to
needs if there isn't hidden usage anymore.

Bug: 137908189
Test: m
Test: check whether device boots successfully
Change-Id: Ida940a2e1bb832cf5c3d147297e7ed65fa6921d5
Merged-In: Ida940a2e1bb832cf5c3d147297e7ed65fa6921d5