Fix widget disappearing issue in migration
In some project, on changing display size, the profile changes.Hence,
need for migration. When migration happens with widgets on workspace,
issue happens.

This CL adds mWidgetId, mProvider in calcDiff() API to compute
differences between src table and dest table for widgets whenever
Migration happens.

The general principle works as -
1.Delete entries from destination table for which appWidgetId,
appWidgetProvider doesn't exist in source table, just before migration.
2.To perform migration, calcDiff() is used to calculate differences of
entries which are present in src table, and not in dest table.

Comparison is done based on mIntent, which is null for widgets.
Hence,mProvider and mappWidgetId parameters which are set for widget,
have been used in calcDiff()

Bug:
Test:manual

Change-Id: I723bfc631746103dd36458e892d717e2dd5b4b12