Fix launcher resuming when starting recents animation
When user clicks on the gesture region of the launcher, this
would start the recents animation, and starting recents animation
could cause launcher resuming, but this would show the launcher
unwanted in some cases.

So we change the condition to start the recents animation, i.e.,
start the recents animation when swiping up and passing the slop.

Change-Id: I49bfa9e021df544aeeb2c7f6eaf26f8f787c8330
Signed-off-by: Jia Jia <jia.jia@zte.com.cn>